requirements.txt
# Backend & API
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.13
prometheus-client==0.17.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.8.6

# SIEM & SOAR clients
loki-client==0.0.1
thehive4py==1.8.0
cortex4py==2.1.0

# Dashboard & Data
jinja2==3.1.3
aiofiles==23.2.1
sqlalchemy==2.0.25
alembic==1.13.1
numpy==1.24.3
pandas==2.0.3
plotly==5.18.0

# Tests
pytest==8.0.0
